refactor(vision): move RingBuffer and its iterator to seperate package, fix RingBuffer clearing, fix angle wrapping calculations in RobotHeadingEstimator with new function, format
