+++ src/main/java/frc/robot/poseestimator/helpers/RingBuffer/RingBufferIterator.java
@@ -1,4 +1,4 @@
-package frc.robot.poseestimator.helpers;
+package frc.robot.poseestimator.helpers.RingBuffer;
 
 import java.util.Iterator;
