+++ src/main/java/frc/utils/AngleUtils.java
@@ -1,0 +1,16 @@
+package frc.utils;
+
+import edu.wpi.first.math.geometry.Rotation2d;
+import frc.robot.constants.MathConstants;
+
+public class AngleUtils {
+
+	public static Rotation2d wrappingAbs(Rotation2d angle) {
+		double rotations = angle.getRotations() % MathConstants.FULL_CIRCLE.getRadians();
+		if (rotations < 0) {
+			rotations += MathConstants.FULL_CIRCLE.getRadians();
+		}
+		return Rotation2d.fromRotations(rotations);
+	}
+
+}
